refactor: drop legacy shims and use canonical imports
