+++ scripts/validate_critical_fix.py
@@ -31,7 +31,7 @@ sys.path.append('.')
 # (mocks removed; see tests/mocks/validate_critical_fix_mocks.py)
 from tests.mocks.validate_critical_fix_mocks import MockOrder, MockContext
 
-from ai_trading.trade_execution import ExecutionEngine
+from ai_trading.execution.engine import ExecutionEngine
 
 def test_production_scenarios():
     """Test the exact scenarios from production logs."""
