--- tests/test_centralized_imports.py
@@ -1,423 +1,0 @@
-"""
-Tests for centralized import management system.
-
-AI-AGENT-REF: Comprehensive tests for dependency import fallbacks
-"""
-
-import pytest
-import sys
-from unittest.mock import patch
-
-
-class TestImportFallbacks:
-    """Test the centralized import system with mocked dependencies."""
-    
-    def setup_method(self):
-        """Clean up any previously imported modules."""
-        # Remove ai_trading.imports from cache if present
-        if 'ai_trading.imports' in sys.modules:
-            del sys.modules['ai_trading.imports']
-    
-    def test_numpy_fallback(self):
-        """Test numpy fallback functionality."""
-        # Test by directly importing - if numpy isn't available, mocks will be used
-        from ai_trading.imports import np
-        
-        # Test that basic operations work regardless of whether numpy is real or mock
-        assert hasattr(np, 'array')
-        assert hasattr(np, 'mean')
-        assert hasattr(np, 'std')
-        assert hasattr(np, 'nan')
-        
-        # Test basic operations
-        arr = np.array([1, 2, 3])
-        assert isinstance(arr, (list, type(np.array([]))))  # Could be list or numpy array
-        mean_result = np.mean([1, 2, 3])
-        assert isinstance(mean_result, (int, float))
-        std_result = np.std([1, 2, 3])
-        assert isinstance(std_result, (int, float))
-    
-    def test_pandas_fallback(self):
-        """Test pandas fallback functionality."""
-        from ai_trading.imports import pd
-        
-        # Test that basic operations work regardless of whether pandas is real or mock
-        assert hasattr(pd, 'DataFrame')
-        assert hasattr(pd, 'Series')
-        assert hasattr(pd, 'read_csv')
-        
-        # Test DataFrame operations
-        df = pd.DataFrame({'A': [1, 2, 3], 'B': [4, 5, 6]})
-        assert hasattr(df, 'shape')
-        assert hasattr(df, 'empty')
-        
-        # Test Series operations
-        series = pd.Series([1, 2, 3])
-        assert hasattr(series, 'mean')
-        mean_result = series.mean()
-        assert isinstance(mean_result, (int, float))
-    
-    def test_sklearn_fallback(self):
-        """Test scikit-learn fallback functionality.""" 
-        from ai_trading.imports import (
-            LinearRegression, StandardScaler, train_test_split,
-            mean_squared_error
-        )
-        
-        # Test LinearRegression
-        lr = LinearRegression()
-        lr.fit([[1, 2], [3, 4]], [1, 2])
-        predictions = lr.predict([[5, 6]])
-        assert hasattr(predictions, '__len__')  # Should be list-like
-        
-        # Test StandardScaler
-        scaler = StandardScaler()
-        scaled = scaler.fit_transform([[1, 2], [3, 4]])
-        assert hasattr(scaled, '__len__')  # Should be list-like
-        
-        # Test train_test_split
-        result = train_test_split([1, 2, 3, 4], [1, 2, 3, 4])
-        assert len(result) == 4  # Should return 4 components
-        
-        # Test metrics
-        mse = mean_squared_error([1, 2], [1, 2])
-        assert isinstance(mse, (int, float))
-    
-    def test_talib_fallback(self):
-        """Test TA-Lib fallback functionality."""
-        from ai_trading.imports import get_ta_lib
-        
-        ta_lib = get_ta_lib()
-        
-        # Test SMA
-        data = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
-        sma = ta_lib.SMA(data, timeperiod=3)
-        assert hasattr(sma, '__len__')  # Should be list-like
-        
-        # Test EMA
-        ema = ta_lib.EMA(data, timeperiod=3)
-        assert hasattr(ema, '__len__')  # Should be list-like
-        
-        # Test RSI
-        rsi = ta_lib.RSI(data, timeperiod=5)
-        assert hasattr(rsi, '__len__')  # Should be list-like
-        
-        # Test MACD
-        macd, signal, histogram = ta_lib.MACD(data)
-        assert all(hasattr(x, '__len__') for x in [macd, signal, histogram])
-        
-        # Test BBANDS
-        upper, middle, lower = ta_lib.BBANDS(data)
-        assert all(hasattr(x, '__len__') for x in [upper, middle, lower])
-    
-    def test_availability_flags(self):
-        """Test that availability flags are properly set."""
-        from ai_trading.imports import (
-            NUMPY_AVAILABLE, PANDAS_AVAILABLE, SKLEARN_AVAILABLE,
-            TALIB_AVAILABLE, PANDAS_TA_AVAILABLE
-        )
-        
-        # These should all be boolean values
-        assert isinstance(NUMPY_AVAILABLE, bool)
-        assert isinstance(PANDAS_AVAILABLE, bool)
-        assert isinstance(SKLEARN_AVAILABLE, bool)
-        assert isinstance(TALIB_AVAILABLE, bool)
-        assert isinstance(PANDAS_TA_AVAILABLE, bool)
-    
-    def test_exports_available(self):
-        """Test that all expected exports are available."""
-        from ai_trading import imports
-        
-        expected_exports = [
-            'np', 'pd', 'sklearn', 'ta', 'talib', 'get_ta_lib',
-            'NUMPY_AVAILABLE', 'PANDAS_AVAILABLE', 'SKLEARN_AVAILABLE',
-            'TALIB_AVAILABLE', 'PANDAS_TA_AVAILABLE',
-            'BaseEstimator', 'TransformerMixin', 'LinearRegression',
-            'RandomForestRegressor', 'StandardScaler', 'train_test_split',
-            'mean_squared_error', 'r2_score'
-        ]
-        
-        for export in expected_exports:
-            assert hasattr(imports, export), f"Missing export: {export}"
-
-
-class TestMockImplementations:
-    """Test specific mock implementation functionality."""
-    
-    def test_mock_numpy_operations(self):
-        """Test MockNumpy mathematical operations."""
-        # AI-AGENT-REF: Fix mock numpy test by directly testing the MockNumpy class
-        # Instead of trying to force the import logic, directly test the mock implementation
-        
-        # Import the MockNumpy class directly from imports module
-        import importlib.util
-        spec = importlib.util.spec_from_file_location("imports", "ai_trading/imports.py")
-        imports_module = importlib.util.module_from_spec(spec)
-        
-        # Temporarily disable numpy import by adding an ImportError
-        original_builtins_import = __builtins__['__import__']
-        
-        def mock_import(name, *args, **kwargs):
-            if name == 'numpy':
-                raise ImportError("Mocked numpy import failure")
-            return original_builtins_import(name, *args, **kwargs)
-        
-        try:
-            # Patch the import function
-            __builtins__['__import__'] = mock_import
-            
-            # Execute the module, which should now use MockNumpy
-            spec.loader.exec_module(imports_module)
-            
-            # Get the numpy object (should be MockNumpy)
-            np = imports_module.np
-            
-            # Verify we got the mock implementation
-            assert hasattr(np, 'array'), "Should have array method"
-            assert not hasattr(np, '__version__'), "Mock numpy should not have __version__"
-            
-            # Test array creation
-            arr = np.array([1, 2, 3, 4, 5])
-            assert arr == [1, 2, 3, 4, 5]
-            
-            # Test zeros and ones
-            zeros = np.zeros(3)
-            assert zeros == [0.0, 0.0, 0.0]
-            
-            ones = np.ones(3)
-            assert ones == [1.0, 1.0, 1.0]
-            
-            # Test mathematical functions
-            assert np.mean([2, 4, 6]) == 4.0
-            assert np.sum([1, 2, 3]) == 6.0
-            assert np.max([1, 5, 3]) == 5.0
-            assert np.min([1, 5, 3]) == 1.0
-            
-            # Test logical operations
-            assert not np.isnan(5.0)
-            assert not np.isinf(5.0)
-            assert np.isfinite(5.0)
-            
-        finally:
-            # Restore original import function
-            __builtins__['__import__'] = original_builtins_import
-    
-    def test_mock_pandas_dataframe(self):
-        """Test MockDataFrame functionality."""
-        # AI-AGENT-REF: Use more targeted import mocking to only block pandas specifically
-        import builtins
-        original_import = builtins.__import__
-        
-        def mock_import(name, *args, **kwargs):
-            if name == 'pandas':
-                raise ImportError(f"No module named '{name}'")
-            return original_import(name, *args, **kwargs)
-        
-        with patch.dict('sys.modules', {'pandas': None}):
-            with patch('builtins.__import__', side_effect=mock_import):
-                # Reload the imports module to trigger fallback
-                import importlib
-                import ai_trading.imports
-                importlib.reload(ai_trading.imports)
-                
-                from ai_trading.imports import pd
-                
-                # Test DataFrame creation with dict
-                df = pd.DataFrame({'A': [1, 2, 3], 'B': [4, 5, 6]})
-                assert not df.empty
-                assert df.shape == (3, 2)
-                
-                # Test column access
-                series_a = df['A']
-                assert len(series_a) == 3
-                assert series_a.data == [1, 2, 3]
-                
-                # Test operations
-                mean_series = df.mean()
-                assert len(mean_series) == 2
-                
-                # Test copy
-                df_copy = df.copy()
-                assert df_copy.shape == df.shape
-    
-    def test_mock_pandas_series(self):
-        """Test MockSeries functionality."""
-        # AI-AGENT-REF: Use more targeted import mocking to only block pandas specifically
-        import builtins
-        original_import = builtins.__import__
-        
-        def mock_import(name, *args, **kwargs):
-            if name == 'pandas':
-                raise ImportError(f"No module named '{name}'")
-            return original_import(name, *args, **kwargs)
-        
-        with patch.dict('sys.modules', {'pandas': None}):
-            with patch('builtins.__import__', side_effect=mock_import):
-                # Reload the imports module to trigger fallback
-                import importlib
-                import ai_trading.imports
-                importlib.reload(ai_trading.imports)
-                
-                from ai_trading.imports import pd
-                
-                # Test Series creation
-                series = pd.Series([1, 2, 3, 4, 5])
-                assert len(series) == 5
-                
-                # Test shift operation
-                shifted = series.shift(1)
-                assert shifted.data[0] is None
-                assert shifted.data[1] == 1
-                
-                # Test rolling window
-                rolling = series.rolling(3)
-                rolling_mean = rolling.mean()
-                assert len(rolling_mean.data) == 5
-                
-                # Test dropna
-                series_with_na = pd.Series([1, None, 3, None, 5])
-                clean_series = series_with_na.dropna()
-                assert None not in clean_series.data
-    
-    def test_mock_sklearn_models(self):
-        """Test MockSklearn model functionality."""
-        # AI-AGENT-REF: Use more targeted sklearn import mocking with proper __builtins__ handling
-        import builtins
-        original_import = builtins.__import__
-        
-        def mock_import(name, *args, **kwargs):
-            if name.startswith('sklearn'):
-                raise ImportError(f"No module named '{name}'")
-            return original_import(name, *args, **kwargs)
-        
-        with patch.dict('sys.modules', {'sklearn': None}):
-            with patch('builtins.__import__', side_effect=mock_import):
-                # Reload the imports module to trigger fallback
-                import importlib
-                import ai_trading.imports
-                importlib.reload(ai_trading.imports)
-                
-                from ai_trading.imports import LinearRegression, RandomForestRegressor, StandardScaler
-                
-                # Test LinearRegression
-                lr = LinearRegression()
-                X = [[1, 2], [3, 4], [5, 6]]
-                y = [1, 2, 3]
-                
-                lr.fit(X, y)
-                predictions = lr.predict(X)
-                assert len(predictions) == len(X)
-                score = lr.score(X, y)
-                assert isinstance(score, float)
-                
-                # Test RandomForestRegressor
-                rf = RandomForestRegressor()
-                rf.fit(X, y)
-                rf_predictions = rf.predict(X)
-                assert len(rf_predictions) == len(X)
-                
-                # Test StandardScaler
-                scaler = StandardScaler()
-                scaled = scaler.fit_transform(X)
-                assert scaled == X  # Mock passes through
-                
-                inverse = scaler.inverse_transform(scaled)
-                assert inverse == X
-    
-    def test_mock_talib_indicators(self):
-        """Test technical indicator functionality."""
-        # Test that indicators work regardless of implementation
-        from ai_trading.imports import talib
-        
-        data = [10, 11, 12, 11, 10, 9, 10, 11, 12, 13, 14, 13, 12, 11, 10]
-        
-        # Test SMA
-        sma = talib.SMA(data, timeperiod=5)
-        assert len(sma) == len(data)
-        assert isinstance(sma[-1], (float, int))  # Last value should be a number
-        
-        # Test EMA
-        ema = talib.EMA(data, timeperiod=5)
-        assert len(ema) == len(data)
-        assert isinstance(ema[-1], (float, int))  # Last value should be a number
-        
-        # Test RSI
-        rsi = talib.RSI(data, timeperiod=5)
-        assert len(rsi) == len(data)
-        
-        # Test MACD
-        macd, signal, histogram = talib.MACD(data)
-        assert len(macd) == len(data)
-        assert len(signal) == len(data)
-        assert len(histogram) == len(data)
-        
-        # Test Bollinger Bands
-        upper, middle, lower = talib.BBANDS(data, timeperiod=5)
-        assert len(upper) == len(data)
-        assert len(middle) == len(data)
-        assert len(lower) == len(data)
-        
-        # Test ATR
-        high = [x + 1 for x in data]
-        low = [x - 1 for x in data]
-        atr = talib.ATR(high, low, data, timeperiod=5)
-        assert len(atr) == len(data)
-        
-        # Test Stochastic
-        slowk, slowd = talib.STOCH(high, low, data)
-        assert len(slowk) == len(data)
-        assert len(slowd) == len(data)
-
-
-class TestRealImports:
-    """Test behavior when real libraries are available."""
-    
-    def test_real_imports_if_available(self):
-        """Test that real imports work when libraries are available."""
-        try:
-            import numpy
-            real_numpy_available = True
-        except ImportError:
-            real_numpy_available = False
-        
-        try:
-            import pandas  
-            real_pandas_available = True
-        except ImportError:
-            real_pandas_available = False
-        
-        from ai_trading.imports import NUMPY_AVAILABLE, PANDAS_AVAILABLE
-        
-        # The availability flags should match what's actually available
-        # (though in test environments they might use mocks)
-        assert isinstance(NUMPY_AVAILABLE, bool)
-        assert isinstance(PANDAS_AVAILABLE, bool)
-    
-    def test_module_imports_work(self):
-        """Test that the module can be imported successfully."""
-        import ai_trading.imports
-        
-        # Basic smoke test - just ensure the module loads
-        assert hasattr(ai_trading.imports, 'np')
-        assert hasattr(ai_trading.imports, 'pd')
-        assert hasattr(ai_trading.imports, 'get_ta_lib')
-
-
-def test_import_summary_logging(caplog):
-    """Test that import summary logging works."""
-    import logging
-    
-    # Enable logging for this test
-    logging.getLogger('ai_trading.imports').setLevel(logging.INFO)
-    
-    # Re-import to trigger logging
-    if 'ai_trading.imports' in sys.modules:
-        del sys.modules['ai_trading.imports']
-    
-    
-    # Check that some logging occurred
-    assert len(caplog.records) >= 0  # May have warnings or info messages
-
-
-if __name__ == '__main__':
-    pytest.main([__file__])
