+++ tests/test_data_fetcher_extended.py
@@ -4,7 +4,7 @@ import types
 import pandas as pd
 import pytest
 
-from ai_trading import data_fetcher
+import ai_trading.data_fetcher as data_fetcher
 
 
 class DummyClient:
