+++ tests/test_meta_learning_module.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import numpy as np
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-from ai_trading import meta_learning
+import ai_trading.meta_learning as meta_learning
 
 
 def test_load_weights_missing(tmp_path, caplog):
